[SIEM] Cypress renaming (#68659)
* updates naming in integration

* updates naming in objects

* updates naming in screens

* updates naming in support

* updates naming in tasks

* fixes failing test
