+++ x-pack/plugins/security_solution/cypress/integration/events_viewer.spec.ts
@@ -32,7 +32,7 @@ import {
   resetFields,
   waitsForEventsToBeLoaded,
 } from '../tasks/hosts/events';
-import { clearSearchBar, kqlSearch } from '../tasks/siem_header';
+import { clearSearchBar, kqlSearch } from '../tasks/security_header';
 
 import { HOSTS_PAGE } from '../urls/navigation';
