+++ x-pack/plugins/security_solution/cypress/integration/navigation.spec.ts
@@ -3,14 +3,14 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-import { DETECTIONS, HOSTS, NETWORK, OVERVIEW, TIMELINES } from '../screens/siem_header';
+import { DETECTIONS, HOSTS, NETWORK, OVERVIEW, TIMELINES } from '../screens/security_header';
 
 import { loginAndWaitForPage } from '../tasks/login';
-import { navigateFromHeaderTo } from '../tasks/siem_header';
+import { navigateFromHeaderTo } from '../tasks/security_header';
 
 import { TIMELINES_PAGE } from '../urls/navigation';
 
-describe('top-level navigation common to all pages in the SIEM app', () => {
+describe('top-level navigation common to all pages in the Security app', () => {
   before(() => {
     loginAndWaitForPage(TIMELINES_PAGE);
   });
