+++ x-pack/plugins/security_solution/cypress/integration/pagination.spec.ts
@@ -12,7 +12,7 @@ import { openAuthentications, openUncommonProcesses } from '../tasks/hosts/main'
 import { waitForUncommonProcessesToBeLoaded } from '../tasks/hosts/uncommon_processes';
 import { loginAndWaitForPage } from '../tasks/login';
 import { goToFirstPage, goToThirdPage } from '../tasks/pagination';
-import { refreshPage } from '../tasks/siem_header';
+import { refreshPage } from '../tasks/security_header';
 
 import { HOSTS_PAGE_TAB_URLS } from '../urls/navigation';
