+++ x-pack/plugins/security_solution/cypress/screens/security_main.ts
