+++ x-pack/plugins/security_solution/cypress/support/commands.js
@@ -30,7 +30,7 @@
 // -- This is will overwrite an existing command --
 // Cypress.Commands.overwrite("visit", (originalFn, url, options) => { ... })
 
-Cypress.Commands.add('stubSIEMapi', function (dataFileName) {
+Cypress.Commands.add('stubSecurityApi', function (dataFileName) {
   cy.on('window:before:load', (win) => {
     // @ts-ignore no null, this is a temp hack see issue above
     win.fetch = null;
