+++ x-pack/plugins/security_solution/cypress/tasks/hosts/uncommon_processes.ts
@@ -5,7 +5,7 @@
  */
 
 import { UNCOMMON_PROCESSES_TABLE } from '../../screens/hosts/uncommon_processes';
-import { REFRESH_BUTTON } from '../../screens/siem_header';
+import { REFRESH_BUTTON } from '../../screens/security_header';
 
 export const waitForUncommonProcessesToBeLoaded = () => {
   cy.get(UNCOMMON_PROCESSES_TABLE).should('exist');
